refactor(design): update DaffSizable interface naming
